Provides option to :ignore_content in equivalence checks
Using the `:ignore_content` option, you may provide one or more
CSS selectors of nodes for which the content should be ignored
when checking for equivalence.

For example, consider my list of devices and Sally's:

Mine:

    <Devices>
      <Device>
        <Name>iPhone</Name>
        <ModelNumber>5</ModelNumber>
        <SerialNumber>1234</SerialNumber>
      </Device>
    </Devices>

Sally's:

    <Devices>
      <Device>
        <Name>iPhone</Name>
        <ModelNumber>5</ModelNumber>
        <SerialNumber>5678</SerialNumber>
      </Device>
    </Devices>

By ignoring the `SerialNumber`, I am able to determine that we
have the same list of devices, since for this purpose I don't
care about a match on the `SerialNumber`.

    EquivalentXml.equivalent?(mine, sallys)
    # => false

    EquivalentXml.equivalent?(mine, sallys, :ignore_content => "SerialNumber")
    # => true